+++ src/app/simulados/components/question-card/question-card.component.ts
@@ -1,16 +1,18 @@
 import { ActivatedRoute } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
-import { QuestionStructure } from '../questoes/question-structure'
-import { QuestionInterface } from '../../utils/model/question-interface';
 import { Component, Input, ViewChild, inject } from '@angular/core';
+import { QuestionInterface } from '../../utils/model/question-interface';
 import { SimuladoEventsService } from '../../utils/service/simulado-events.service';
 import { DinamicLoaderDirective } from '../../utils/directive/dinamic-loader.directive';
-import { SelectQuestionComponent } from '../questoes/select-question/select-question.component';
-import { OptionsQuestionComponent } from '../questoes/options-question/options-question.component';
-import { MultipleQuestionComponent } from '../questoes/multiple-question/multiple-question.component';
-import { DragDropQuestionComponent } from '../questoes/drag-drop-question/drag-drop-question.component';
 import { ReportDialogComponent } from '../../../shared/components/report-dialog/report-dialog.component';
-import { TrueFalseQuestionComponent } from '../questoes/true-false-question/true-false-question.component';
+import {
+  QuestionStructure,
+  DragDropQuestionComponent,
+  MultipleQuestionComponent,
+  OptionsQuestionComponent,
+  SelectQuestionComponent,
+  TrueFalseQuestionComponent
+} from '../questions';
 
 @Component({
   selector: 'app-question-card',
@@ -18,7 +20,7 @@ import { TrueFalseQuestionComponent } from '../questoes/true-false-question/true
   styleUrl: './question-card.component.css'
 })
 export class QuestionCardComponent {
-  @ViewChild(DinamicLoaderDirective) private dynamicHost!: DinamicLoaderDirective; 
+  @ViewChild(DinamicLoaderDirective) private dynamicHost!: DinamicLoaderDirective;
   @Input() questionData: QuestionInterface;
   @Input() questionIndex: number;
   @Input() totalQuestions: number;
@@ -27,31 +29,31 @@ export class QuestionCardComponent {
   questionInstance: QuestionStructure;
   readonly dialog = inject(MatDialog);
 
-  constructor(private route: ActivatedRoute){
+  constructor(private route: ActivatedRoute) {
     this.examName = this.route.snapshot.paramMap.get("exam");
   }
-  
+
   ngAfterViewInit() {
     this.loadQuestionComponent();
   }
-  
+
   /* Updates the question data and reload the question type component */
-  public setQuestionData(data: QuestionInterface){
+  public setQuestionData(data: QuestionInterface) {
     this.questionData = data;
     this.loadQuestionComponent();
   }
-  
+
   /* Initializes the question structure component */
   loadQuestionComponent() {
     this.questionType = this.questionData.type;
 
     /* Randomize options array */
-    if(this.questionType != "truefalse" && this.questionType != "dragdrop"){
+    if (this.questionType != "truefalse" && this.questionType != "dragdrop") {
       this.questionData.options = this.questionData.options.sort(() => Math.random() - Math.random());
     }
     /* Clears the question structure component */
     this.dynamicHost.view.clear();
-    
+
     /* Set the component based on the question type */
     const questionComponentMap = {
       'multiple': MultipleQuestionComponent,
@@ -61,7 +63,7 @@ export class QuestionCardComponent {
       'dragdrop': DragDropQuestionComponent
     };
     const questionComponentType = questionComponentMap[this.questionType];
-    
+
     /* Creates the question structure component */
     const viewContainerRef = this.dynamicHost.view;
     this.questionInstance = viewContainerRef.createComponent<QuestionStructure>(questionComponentType).instance;
@@ -73,12 +75,12 @@ export class QuestionCardComponent {
   /* Opens dialog window for reporting */
   reportQuestion() {
     this.dialog.open(ReportDialogComponent, {
-      data: "Question "+this.questionData.id+": content/"+this.examName+"/questions.json"
+      data: "Question " + this.questionData.id + ": content/" + this.examName + "/questions.json"
     });
   }
 
   /* Finishes the exam */
-  finishExam(){
+  finishExam() {
     SimuladoEventsService.get('endExam').emit(this.questionInstance.getSummary());
   }
 }
