+++ src/app/simulados/components/questions/index.ts
@@ -1,0 +1,6 @@
+export { QuestionStructure } from "./question-structure";
+export { DragDropQuestionComponent } from "./drag-drop-question/drag-drop-question.component";
+export { MultipleQuestionComponent } from "./multiple-question/multiple-question.component";
+export { OptionsQuestionComponent } from "./options-question/options-question.component";
+export { SelectQuestionComponent } from "./select-question/select-question.component";
+export { TrueFalseQuestionComponent } from "./true-false-question/true-false-question.component";
