+++ src/app/simulados/components/questions/true-false-question/true-false-question.component.ts
